fix path check for mac

+++ launcher/game/new_project.rpy
@@ -21,6 +21,7 @@
 
 init python:
     import os
+    import shutil
     from extractor import Extractor
 
     extract = Extractor()
@@ -102,22 +103,23 @@ label add_a_mod:
                 interface.error(_("The operation has been cancelled."))
                 renpy.jump("front_page")
 
-            if path.endswith('.zip'):
-                valid = extract.valid_zip(path)
-                if valid:
-                    pass
+            if not persistent.safari or not renpy.macintosh:
+                if path.endswith('.zip'):
+                    valid = extract.valid_zip(path)
+                    if valid:
+                        pass
+                    else:
+                        shutil.rmtree(project_dir)
+                        inteface.error(_("The mod ZIP you selected is not a valid DDLC mod archive.\nSelect a different mod ZIP and try again."),)
+                        renpy.jump("front_page")
+                elif path.endswith('.rar'):
+                    shutil.rmtree(project_dir)
+                    inteface.error(_("RAR files cannot be unzipped or unrarred by DDML.\nConvert the file to a ZIP file and try again."),)
+                    renpy.jump("front_page")
                 else:
                     shutil.rmtree(project_dir)
-                    inteface.error(_("The mod ZIP you selected is not a valid DDLC mod archive.\nSelect a different mod ZIP and try again."),)
+                    inteface.error(_("Unknown file type.\nSelect a DDMC mod ZIP file and try again."),)
                     renpy.jump("front_page")
-            elif path.endswith('.rar'):
-                shutil.rmtree(project_dir)
-                inteface.error(_("RAR files cannot be unzipped or unrarred by DDML.\nConvert the file to a ZIP file and try again."),)
-                renpy.jump("front_page")
-            else:
-                shutil.rmtree(project_dir)
-                inteface.error(_("Unknown file type.\nSelect a DDMC mod ZIP file and try again."),)
-                renpy.jump("front_page")
 
             interface.interaction(_("Installing the Mod"), _("This process may take some time. Please wait."),)
